refactor: Implement useRequestPasswordChange hook
